add `make plugin` target
it compiles & tests the current source code and creates a new
distribution directory

    geocouch-$(GEOCOUCH_VERSION)-$(ERLANG_VERSION)

and a corresponding .tar.gz file:

    geocouch-$(GEOCOUCH_VERSION)-$(ERLANG_VERSION).tar.gz

and produces a sha hash as produced by Erlang’s `crypto:sha()`
function.

The resulting tarball can be published on anywhere on HTTP
together with the sha.

This whole shebang then can be installed with, e.g.:

    couch_plugins:install({
      "geocouch",
       "http://people.apache.org/~jan/",
       "couchdb1.2.x_v0.3.0-11-gd83ba22",
       [{"R15B03", "ZetgdHj2bY2w37buulWVf3USOZs="}]}).

(In this case, the tarball had been uploaded to
 http://people.apache.org/~jan/)